+++ packages/vue-form-control/src/composables/index.ts
@@ -11,4 +11,5 @@ export * from './date-input';
 export * from './imask';
 export * from './wrapper';
 export * from './group';
+export * from './set';
 export * from './form';